py3k: long to int

+++ src/collectors/interrupt/interrupt.py
@@ -13,7 +13,6 @@ The InterruptCollector class collects metrics on interrupts from
 import platform
 import os
 import diamond.collector
-from diamond.pycompat import long
 
 # Detect the architecture of the system
 # and set the counters for MAX_VALUES
@@ -69,7 +68,7 @@ class InterruptCollector(diamond.collector.Collector):
                     metric_value = data[1]
                     self.publish(metric_name,
                                  self.derivative(metric_name,
-                                                 long(metric_value),
+                                                 int(metric_value),
                                                  counter))
                 else:
                     if len(data[0]) == cpuCount + 1:
@@ -91,7 +90,7 @@ class InterruptCollector(diamond.collector.Collector):
 
                         metric_name_node = metric_name + 'CPU' + str(index - 1)
                         value = int(self.derivative(metric_name_node,
-                                                    long(value), counter))
+                                                    int(value), counter))
                         total += value
                         self.publish(metric_name_node, value)
